Add support for cc_on_error parameter.
cc_on_error allows individual repositories to specify via query
parameter specific teams or individuals who should be cc'd in all cases
in the event of an error. I think a query parameter is the way to go
since it allows the value to be changed by anyone with admin access to
the repository being published rather than someone with code or
environment variable control access to publisher itself.